requirements.txt
pygame
websocket-client
numpy
ccxt
pandas
ta
dydx-v3-python